Add direct feedback on text check modal
RISDEV-7553
